Backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Pydantic
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Firebase (NUEVO)
firebase-admin>=6.4.0

# Google Cloud
google-cloud-storage>=2.14.0
google-cloud-tasks>=2.15.0
google-generativeai>=0.8.0

# ML/NLP
transformers>=4.36.2
torch>=2.1.2
sentencepiece>=0.1.99

# PDF/Document Processing
PyPDF2>=3.0.1
pdfplumber>=0.10.0
python-docx>=1.1.0
PyMuPDF>=1.23.0
Pillow>=10.0.0

# Utilities
requests>=2.31.0
thefuzz>=0.20.0
python-Levenshtein>=0.23.0
reportlab>=4.0.9
python-multipart>=0.0.6
python-dotenv>=1.0.0
